refactor: 입출력 부분을 제외한 로직을 BowlingGame으로 이동
