+++ src/main/java/bowling/controller/BowlingController.java
@@ -1,11 +1,9 @@
 package bowling.controller;
 
-import bowling.domain.frame.RoundNumber;
 import bowling.domain.pin.Pin;
 import bowling.domain.player.Player;
 import bowling.domain.player.PlayerName;
 import bowling.exception.BowlingException;
-import bowling.view.BowlingGame;
 import bowling.view.InputView;
 import bowling.view.OutputView;
 
@@ -28,37 +26,12 @@ public final class BowlingController {
     public void run() {
         final int playerCount = inputView.inputPlayerCount();
         final List<PlayerName> playerNames = inputView.inputPlayerNames(playerCount);
-
         final BowlingGame bowlingGame = new BowlingGame(playerNames);
         outputView.printScoreBoard(bowlingGame.scoreBoard());
 
-        for (int i = RoundNumber.MIN; i <= RoundNumber.MAX; i++) {
-            bowlAll(bowlingGame, i);
-        }
-    }
-
-    private void bowlAll(BowlingGame bowlingGame, int roundNumber) {
-        for (Player player : bowlingGame.players()) {
-            bowl(bowlingGame, roundNumber, player);
-        }
-    }
-
-    private void bowl(BowlingGame bowlingGame, int roundNumberSource, Player player) {
-        final RoundNumber roundNumber = new RoundNumber(roundNumberSource);
         final BowlingTemplate bowlingTemplate =
                 new BowlingTemplate(this::pin, () -> outputView.printScoreBoard(bowlingGame.scoreBoard()));
-        while (!player.isEnded(roundNumber)) {
-            downPin(player, roundNumber, bowlingTemplate);
-        }
-    }
-
-    private void downPin(final Player player, final RoundNumber roundNumber, final BowlingTemplate bowlingTemplate) {
-        try {
-            bowlingTemplate.execute(player, pin -> player.knockDownPin(roundNumber, pin));
-        } catch (BowlingException e) {
-            System.err.println(e.getMessage());
-            downPin(player, roundNumber, bowlingTemplate);
-        }
+        bowlingGame.playGame(bowlingTemplate);
     }
 
     private Pin pin(Player player) {
